split `processing` module
